[clangd] Fix stack trace may be omitted when clangd is crashed.
We used PrintStackTraceOnErrorSignals to emit stack trace to llvm::errs() when clangd receives an abort/kill signal (e.g. an assertion failure), but we set llvm::errs() as buffered, which may cause the stack trace is not emitted to standard error.

To fix it, we implement our own signal handler which flushs llvm::errs()
immediately.

TESTED

run "./bin/clangd < crash_input.json 2> error.log", and verify the stack
trace is in `error.log`.